+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Adds `VectorClusterGroup` concept to the `VcsApp`. For more details see [the documentation](documentation/vectorClusterGroup.md).
 - The `FeatureAtPixelInteraction` has been extended by a drill pick API. The `InteractionEvent` now has a `features` property along
   with the `feature` property. Drill pick will return all features (to a certain depth) at the pixel, not just the topmost one.
+- Adds a `resetExclusiveMapControls` in the mapCollection to reset the exclusiveMapControls.
 
 ### 6.0.6